+++ alura-exercises/Python/ATM/utils.py
@@ -1,0 +1,6 @@
+def check_y_n(question):
+  answer = input(f'{question} [y/n] ')
+  while (answer.lower() != "y" and answer.lower() != "n"):
+    print ("Opção inválida.")
+    answer = input(f'{question} [y/n] ')
+  return answer
